requirements.txt
pandas
setuptools
pyyaml
typing-extensions
pathlib
sentence-transformers
numpy
scikit-learn
python-dotenv
langchain-core
langchain-groq
groq
fastapi